% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codacore.R
\name{activeInputs.codacore}
\alias{activeInputs.codacore}
\title{activeInputs}
\usage{
activeInputs.codacore(cdcr)
}
\arguments{
\item{cdcr}{A codacore object.}
}
\value{
The covariates included in the log-ratios
}
\description{
activeInputs
}
