% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codacore.R
\name{getBinaryPartitions}
\alias{getBinaryPartitions}
\title{getBinaryPartitions}
\usage{
getBinaryPartitions(cdcr)
}
\arguments{
\item{cdcr}{A codacore object}
}
\value{
A matrix describing whether each component (as rows) is found in the
 numerator (1) or denominator (-1) of each learned log-ratio (as columns).
 This format resembles a serial binary partition matrix frequently used
 in balance analysis.
}
\description{
getBinaryPartitions
}
