% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codacore.R
\name{getLogRatios}
\alias{getLogRatios}
\title{getLogRatios}
\usage{
getLogRatios(cdb, x = NULL)
}
\arguments{
\item{cdb}{A codacore object}

\item{x}{A set of (possibly unseen) compositional data. 
The covariates must be passed in the same order as 
for the original codacore() call.}
}
\value{
The learned log-ratio features, computed on input x.
}
\description{
getLogRatios
}
