% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codacore.R
\name{getNumeratorParts}
\alias{getNumeratorParts}
\title{getNumeratorParts}
\usage{
getNumeratorParts(cdb, baseLearnerIndex = 1)
}
\arguments{
\item{cdb}{A codacore object.}

\item{baseLearnerIndex}{An integer indicating which of the 
(possibly multiple) log-ratios learned by codacore to be used.}
}
\value{
The covariates in the numerator of the selected log-ratio.
}
\description{
getNumeratorParts
}
