% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codacore.R
\name{getNumLogRatios}
\alias{getNumLogRatios}
\title{getNumLogRatios}
\usage{
getNumLogRatios(cdcr)
}
\arguments{
\item{cdcr}{A codacore object}
}
\value{
The number of log-ratios that codacore found.
    Typically a small integer. Can be zero if codacore
    found no predictive log-ratios in the data.
}
\description{
getNumLogRatios
}
