% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codacore.R
\name{getSlopes}
\alias{getSlopes}
\title{getSlopes}
\usage{
getSlopes(cdcr)
}
\arguments{
\item{cdcr}{A codacore object}
}
\value{
The slopes (i.e., regression coefficients) for each log-ratio.
}
\description{
getSlopes
}
