% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codacore.R
\name{getTidyTable}
\alias{getTidyTable}
\title{getTidyTable}
\usage{
getTidyTable(cdcr)
}
\arguments{
\item{cdcr}{A codacore object}
}
\value{
A table displaying the log-ratios found.
}
\description{
getTidyTable
}
