% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codacore.R
\name{predict.codacore}
\alias{predict.codacore}
\title{predict}
\usage{
\method{predict}{codacore}(object, x, logits = T)
}
\arguments{
\item{object}{A codacore object.}

\item{x}{A set of inputs to our model.}

\item{logits}{Whether to return outputs in logit space
(as opposed to probability space). Should always be set
to TRUE for regression with continuous outputs, but can
be toggled for classification problems.}
}
\description{
predict
}
