% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{simulateHTS}
\alias{simulateHTS}
\title{simulateHTS}
\usage{
simulateHTS(n, p, outputType = "binary", logratio = "simple")
}
\arguments{
\item{n}{Number of observations.}

\item{p}{Number of covariates.}

\item{outputType}{A string indicating 'binary' or 'continuous'.}

\item{logratio}{A string indicating 'simple', 'balance', or 
'amalgamation'.}
}
\value{
A list containing a matrix of inputs and a vector of outputs
}
\description{
This function simulates a set of (x, y) pairs.
The covariates x are compositional, meaning they only
carry relative information.
The response y is a binary indicator.
The rule linking x and y can be a balance or an amalgamation.
}
