% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codacore.R
\name{codacore}
\alias{codacore}
\title{codacore}
\usage{
codacore(
  x,
  y,
  type = "balances",
  mode = "classification",
  lambda = 1,
  shrinkage = 1,
  maxBaseLearners = 10,
  optParams = list(),
  cvParams = list(),
  verbose = F,
  overlap = T
)
}
\arguments{
\item{x}{A data.frame of the compositional predictor variables.}

\item{y}{A data.frame of the response variables.}

\item{type}{A string indicating whether to use "balances" or "amalgamations".
Also accepts "balance", "B", "ILR", or "amalgam", "A", "SLR".}

\item{mode}{A string indicating "classification" or "regression".}

\item{lambda}{A numeric. Corresponds to the "lambda-SE" rule. Sets the "regularization strength"
used by the algorithm to decide how to harden the ratio. 
Larger numbers tend to yield fewer, more sparse ratios.}

\item{shrinkage}{A numeric. Shrinkage factor applied to each base learner.
Defaults to 1.0, i.e., no shrinkage applied.}

\item{maxBaseLearners}{An integer. The maximum number of log-ratios that the model will
learn before stopping. Automatic stopping based on \code{seRule} may occur sooner.}

\item{optParams}{A list of named parameters for the optimization of the
continuous relaxation. Empty by default. User can override as few or as
many of our defaults as desired. Includes adaptiveLR (learning rate under
adaptive training scheme), momentum (in the gradient-descent sense), 
epochs (number of gradient-descent epochs), batchSize (number of 
observations per minibatch, by default the entire dataset),
and vanillaLR (the learning rate to be used if the user does *not* want
to use the 'adaptiveLR', to be used at the risk of optimization issues).}

\item{cvParams}{A list of named parameters for the "hardening" procedure
using cross-validation. Includes numFolds (number of folds) and
maxCutoffs (number of candidate cutoff values of 'c' to be tested out
during CV process).}

\item{verbose}{A logical. Toggles whether to display intermediate steps.}

\item{overlap}{TODO: To be implemented}
}
\description{
This function implements the codacore algorithm described in [DOI].
}
