% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codacore.R
\name{plot.codacore}
\alias{plot.codacore}
\title{plot}
\usage{
\method{plot}{codacore}(x, ...)
}
\arguments{
\item{x}{A codacore object.}

\item{...}{Not used.}
}
\description{
Plots a summary of a fitted codacore model.
Credit to the authors of the selbal package (Rivera-Pinto et al., 2018),
from whose package these plots were inspired.
}
